Change adjustReserve type from int32 to int8

+++ src/libxrpl/tx/transactors/oracle/SetOracle.cpp
@@ -92,7 +92,7 @@ SetOracle::preclaim(PreclaimContext const& ctx)
         return !v || *v == (*sle)[field];
     };
 
-    std::int32_t adjustReserve = 0;
+    std::int8_t adjustReserve = 0;
     if (sle)
     {
         // update
